\name{toydata-package}
\alias{toydata-package}
\alias{toydata}
\docType{package}
\title{
\packageTitle{toydata}
}
\description{
\packageDescription{toydata}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{toydata}
\packageIndices{toydata}
Data manipulation
}
\author{
\packageAuthor{toydata}

Maintainer: \packageMaintainer{toydata}
}

\keyword{ package }

